move: use-bookmark-query 파일 위치 이동, any 타입 수정 (#164)
